fix int conversion for ess

+++ pyabc/epsilon/temperature.py
@@ -459,8 +459,8 @@ class EssScheme(TemperatureScheme):
         # execute functino (expensive if in calibration)
         df = get_weighted_distances()
 
-        weights = np.array(df['w'])
-        pdfs = np.array(df['distance'])
+        weights = np.array(df['w'], dtype=float)
+        pdfs = np.array(df['distance'], dtype=float)
 
         # compute rescaled posterior densities
         if kernel_scale == SCALE_LIN:
